Recreated the raycaster to support lines instead of rects and generally work better
Changed enviornment to have lines instead of rects as walls. These are just generally better because they can be angled

Added proper net support, and played with formatting of the screen to be able to see collisions better

Next steps: implement closest hole values as well as the actual hole functions. After will be wrapping the simulation into a proper package so that I can begin training the network