Added hole game mechanic, including a timer for the hunter to check.
Added a global frame counter that automatically ends the simulation after 1000 frames
	Also ends if all players captured

added capture mechanics and fitness of both players and hunters

Next Step is really just to turn into a function, get the fitnesses worst player fitness and hunter fitness while also getting the average certainty of all true functions
This would then be used as the rate for which to shuffle the next generation. This assumes that there is an inherent link between certainty and correctness (not neccessarily true but its good enough). This will make more drastic if its not certain what to do